+++ wrappers/LikesWrapper.jsx
@@ -1,7 +1,0 @@
-"use client";
-
-import { LikesProvider } from "@/context/LikesContext";
-
-export default function LikesWrapper({ children }) {
-  return <LikesProvider>{children}</LikesProvider>;
-}